+++ packages/assets-controllers/src/NftDetectionController.test.ts
@@ -4,6 +4,7 @@ import {
   ChainId,
   InfuraNetworkType,
 } from '@metamask/controller-utils';
+import { MockInternalProvider } from '@metamask/eth-json-rpc-provider';
 import { MOCK_ANY_NAMESPACE, Messenger } from '@metamask/messenger';
 import type {
   MessengerActions,
@@ -26,7 +27,6 @@ import type { PreferencesState } from '@metamask/preferences-controller';
 import nock from 'nock';
 
 import { FakeBlockTracker } from '../../../tests/fake-block-tracker.js';
-import { FakeProvider } from '../../../tests/fake-provider.js';
 import { jestAdvanceTime } from '../../../tests/helpers.js';
 import { createMockInternalAccount } from '../../accounts-controller/tests/mocks.js';
 import {
@@ -771,7 +771,7 @@ describe('NftDetectionController', () => {
 
   it('should return true if mainnet is detected', async () => {
     const mockAddNfts = jest.fn();
-    const provider = new FakeProvider();
+    const provider = new MockInternalProvider();
     const mockNetworkClient: NetworkClient = {
       configuration: {
         chainId: ChainId.mainnet,