+++ packages/wallet/src/initialization/instances/transaction-controller/transaction-controller.ts
@@ -47,7 +47,7 @@ export const transactionController: InitializationConfiguration<
       events: [
         'AccountActivityService:transactionUpdated',
         // TODO: Replace with `NetworkController:stateChanged` once TransactionController migrates.
-        // eslint-disable-next-line no-restricted-syntax
+
         'NetworkController:stateChange',
       ],
     });